+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix deserialization for empty JSON responses
 - Fixed bug in query services where lists had incorrect parent item in request
 - Improve deserializer of XML error responses
+- Adds Serverless Repo service
 
 ## [0.32.0] - 2018-03-03